Fixing compile issues and merging into build process

+++ canny/labs/filters.h
@@ -2,7 +2,7 @@
 #define __FILTERS_H__
 #define FILTERSIZE 3
 
-void Conv2DSerial(int *inImg, int *outImg, double filter[FILTERSIZE][FILTERSIZE], int width, int height, int filterSize); 
+void Conv2DSerial(float *inImg, float *outImg, double filter[FILTERSIZE][FILTERSIZE], int width, int height, int filterSize);
 void populate_blur_filter(double outFilter[FILTERSIZE][FILTERSIZE]);
 void GradientSobelSerial(int *inImg, float *sobelImg, float *gradientImg, int height, int width); 
 #ifdef __CUDACC__
